Add convert hook
