+++ src/components/markdown-renderer/basic-markdown-renderer.tsx
@@ -5,16 +5,14 @@
  */
 
 import MarkdownIt from 'markdown-it'
-import React, { ReactElement, RefObject, useMemo, useRef } from 'react'
-import ReactHtmlParser from 'react-html-parser'
+import React, { RefObject, useMemo } from 'react'
 import { useSelector } from 'react-redux'
 import { ApplicationState } from '../../redux'
 import { DocumentLengthLimitReachedAlert } from './document-length-limit-reached-alert'
+import { useConvertMarkdownToReactDom } from './hooks/use-convert-markdown-to-react-dom'
 import './markdown-renderer.scss'
 import { ComponentReplacer } from './replace-components/ComponentReplacer'
-import { AdditionalMarkdownRendererProps, LineKeys } from './types'
-import { buildTransformer } from './utils/html-react-transformer'
-import { calculateNewLineNumberMapping } from './utils/line-number-mapping'
+import { AdditionalMarkdownRendererProps } from './types'
 
 export interface BasicMarkdownRendererProps {
   componentReplacers?: () => ComponentReplacer[],
@@ -36,30 +34,8 @@ export const BasicMarkdownRenderer: React.FC<BasicMarkdownRendererProps & Additi
     onPostRendering
   }) => {
   const maxLength = useSelector((state: ApplicationState) => state.config.maxDocumentLength)
-
-  const oldMarkdownLineKeys = useRef<LineKeys[]>()
-  const lastUsedLineId = useRef<number>(0)
-
-  const markdownReactDom: ReactElement[] = useMemo(() => {
-    if (onPreRendering) {
-      onPreRendering()
-    }
-    const trimmedContent = content.substr(0, maxLength)
-    const html: string = markdownIt.render(trimmedContent)
-    const contentLines = trimmedContent.split('\n')
-    const {
-      lines: newLines,
-      lastUsedLineId: newLastUsedLineId
-    } = calculateNewLineNumberMapping(contentLines, oldMarkdownLineKeys.current ?? [], lastUsedLineId.current)
-    oldMarkdownLineKeys.current = newLines
-    lastUsedLineId.current = newLastUsedLineId
-    const transformer = componentReplacers ? buildTransformer(newLines, componentReplacers()) : undefined
-    const rendering = ReactHtmlParser(html, { transform: transformer })
-    if (onPostRendering) {
-      onPostRendering()
-    }
-    return rendering
-  }, [onPreRendering, onPostRendering, content, maxLength, markdownIt, componentReplacers])
+  const trimmedContent = useMemo(() => content.length > maxLength ? content.substr(0, maxLength) : content, [content, maxLength])
+  const markdownReactDom = useConvertMarkdownToReactDom(trimmedContent, markdownIt, componentReplacers, onPreRendering, onPostRendering)
 
   return (
     <div className={`${className ?? ''} d-flex flex-column align-items-center ${wide ? 'wider' : ''}`}>
